+++ PresentationLayer/UIComponents/Screens/DeviceInfo/DeviceInfoFactory.swift
@@ -1,0 +1,427 @@
+//
+//  DeviceInfoFactory.swift
+//  wxm-ios
+//
+//  Created by Pantelis Giazitsis on 9/10/24.
+//
+
+import Foundation
+import DomainLayer
+import SwiftUI
+import Toolkit
+
+extension DeviceInfoViewModel.Field {
+	var warning: DeviceInfoRowView.Row.Warning? {
+		switch self {
+			case .remove:
+				return .desructive(LocalizableString.deviceInfoStationRemoveWarning.localized)
+			default:
+				return nil
+		}
+	}
+
+	func titleFor(devie: DeviceDetails) -> String {
+		switch self {
+			case .name:
+				return LocalizableString.deviceInfoStationName.localized
+			case .frequency:
+				return LocalizableString.deviceInfoStationFrequency.localized
+			case .reboot:
+				return LocalizableString.deviceInfoStationReboot.localized
+			case .maintenance:
+#warning("TODO: Format when info is ready")
+				return LocalizableString.deviceInfoStationMaintenance("").localized
+			case .remove:
+				return LocalizableString.deviceInfoStationRemove.localized
+			case .reconfigureWifi:
+				return LocalizableString.deviceInfoStationReconfigureWifi.localized
+			case .stationLocation:
+				return LocalizableString.deviceInfoStationLocation.localized
+			case .rewardSplit:
+				return LocalizableString.RewardDetails.rewardSplit.localized
+		}
+	}
+
+	func descriptionFor(device: DeviceDetails, for followState: UserDeviceFollowState?, deviceInfo: NetworkDevicesInfoResponse?) -> String {
+		switch self {
+			case .name:
+				return device.displayName
+			case .frequency:
+				if device.isHelium {
+					return LocalizableString.deviceInfoStationHeliumFrequencyDescription(DisplayedLinks.heliumRegionFrequencies.linkURL).localized
+				}
+
+				return ""
+			case .reboot:
+				return LocalizableString.deviceInfoStationRebootDescription.localized
+			case .maintenance:
+				return  LocalizableString.deviceInfoStationMaintenanceDescription(DisplayedLinks.heliumTroubleshooting.linkURL).localized
+			case .remove:
+				return LocalizableString.deviceInfoStationRemoveDescription(DisplayedLinks.documentationLink.linkURL).localized
+			case .reconfigureWifi:
+				return LocalizableString.deviceInfoStationReconfigureWifiDescription.localized
+			case .stationLocation:
+				if followState?.relation == .owned {
+					return LocalizableString.deviceInfoOwnedStationLocationDescription(device.address ?? "").localized
+				}
+				return LocalizableString.deviceInfoStationLocationDescription(device.address ?? "").localized
+			case .rewardSplit:
+				guard let rewardSplit = deviceInfo?.rewardSplit else {
+					return ""
+				}
+				let count = rewardSplit.count
+				return LocalizableString.RewardDetails.rewardSplitDescription(count).localized
+		}
+	}
+
+	func imageUrlFor(device: DeviceDetails, followState: UserDeviceFollowState?) -> URL? {
+		switch self {
+			case .name:
+				return nil
+			case .frequency:
+				return nil
+			case .reboot:
+				return nil
+			case .maintenance:
+				return nil
+			case .remove:
+				return nil
+			case .reconfigureWifi:
+				return nil
+			case .stationLocation:
+				guard let cellCenterLocation = device.cellCenter?.toCLLocationCoordinate2D() else {
+					return nil
+				}
+
+				let isOwned = followState?.relation == .owned
+				let marker = isOwned ? device.location?.toCLLocationCoordinate2D() : nil
+				let polygon = device.cellPolygon?.map { $0.toCLLocationCoordinate2D() }
+				let options = MapBoxSnapshotUrlGenerator.Options(location: cellCenterLocation,
+															markerLocation: marker,
+															size: MapBoxConstants.snapshotSize,
+															zoomLevel: MapBoxConstants.snapshotZoom,
+															polygon: polygon)
+				let fetcher = MapBoxSnapshotUrlGenerator(options: options)
+
+				return fetcher.getUrl()
+			case .rewardSplit:
+				return nil
+		}
+	}
+
+	func buttonInfoFor(devie: DeviceDetails, followState: UserDeviceFollowState?) -> DeviceInfoButtonInfo? {
+		switch self {
+			case .name:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonChangeName.localized)
+			case .frequency:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonChangeFrequency.localized)
+			case .reboot:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonReboot.localized)
+			case .maintenance:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonEnterMaintenance.localized)
+			case .remove:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonRemove.localized)
+			case .reconfigureWifi:
+				return .init(icon: nil, title: LocalizableString.deviceInfoButtonReconfigureWifi.localized)
+			case .stationLocation:
+				guard followState?.relation == .owned else {
+					return nil
+				}
+				return .init(icon: .editIcon,
+							 title: LocalizableString.deviceinfoStationLocationButtonTitle.localized,
+							 buttonStyle: .filled())
+			case .rewardSplit:
+				return nil
+		}
+	}
+
+	func customViewFor(deviceInfo: NetworkDevicesInfoResponse?) -> AnyView? {
+		switch self {
+			case .rewardSplit:
+				guard let rewardsplit = deviceInfo?.rewardSplit, rewardsplit.count > 0 else {
+					return nil
+				}
+				let items = rewardsplit.map { split in
+					let userWallet = MainScreenViewModel.shared.userInfo?.wallet?.address
+					let isUserWallet = split.wallet == userWallet
+					return split.toSplitViewItem(showReward: false, isUserWallet: isUserWallet)
+				}
+				return RewardsSplitView.WalletsListView(items: items).toAnyView
+			default:
+				return nil
+		}
+	}
+}
+
+extension DeviceInfoViewModel.InfoField {
+	static func getShareText(for device: DeviceDetails, deviceInfo: NetworkDevicesInfoResponse?, mainVM: MainScreenViewModel, followState: UserDeviceFollowState?) -> String {
+		var fields: [Self] = []
+		if device.isHelium {
+			fields = heliumFields
+		} else {
+			fields = wifiFields
+		}
+
+		let textComps: [String] = fields.compactMap { field in
+			guard let value = field.value(for: device, deviceInfo: deviceInfo, mainVM: mainVM, followState: followState) else {
+				return nil
+			}
+			return "\(field.title): \(value)"
+		}
+
+		return textComps.joined(separator: "\n")
+	}
+
+	private static func getRssiText(rssi: String?, lastActivityDate: Date?, timezone: String?) -> String? {
+		guard var rssi else {
+			return nil
+		}
+		rssi = "\(rssi) \(UnitConstants.DBM)"
+		let timestamp = lastActivityDate?.getFormattedDate(format: .monthLiteralYearDayTime,
+														   timezone: timezone?.toTimezone ?? .current).capitalizedSentence
+		let elements = [rssi, timestamp].compactMap { $0 }
+		return elements.joined(separator: " @ ")
+	}
+
+	var title: String {
+		switch self {
+			case .name:
+				return LocalizableString.deviceInfoStationInfoName.localized
+			case .bundleName:
+				return LocalizableString.deviceInfoStationInfoBundleName.localized
+			case .gatewayModel, .stationModel:
+				return LocalizableString.deviceInfoStationInfoModel.localized
+			case .devEUI:
+				return LocalizableString.deviceInfoStationInfoDevEUI.localized
+			case .hardwareVersion:
+				return LocalizableString.deviceInfoStationInfoHardwareVersion.localized
+			case .firmwareVersion:
+				return LocalizableString.deviceInfoStationInfoFirmwareVersion.localized
+			case .lastHotspot:
+				return LocalizableString.deviceInfoStationInfoLastHotspot.localized
+			case .lastRSSI:
+				return LocalizableString.deviceInfoStationInfoLastRSSI.localized
+			case .serialNumber:
+				return LocalizableString.deviceInfoStationInfoSerialNumber.localized
+			case .ATECC:
+				return LocalizableString.deviceInfoStationInfoATECC.localized
+			case .GPS:
+				return LocalizableString.deviceInfoStationInfoGPS.localized
+			case .wifiSignal:
+				return LocalizableString.deviceInfoStationInfoWifiSignal.localized
+			case .batteryState:
+				return LocalizableString.deviceInfoStationInfoBattery.localized
+			case .claimedAt:
+				return LocalizableString.deviceInfoClaimDate.localized
+			case .lastGatewayActivity:
+				return LocalizableString.deviceInfoLastGatewayActivity.localized
+			case .lastStationActivity:
+				return LocalizableString.deviceInfoLastStationActivity.localized
+			case .stationRssi:
+				return LocalizableString.deviceInfoStationRssi.localized
+		}
+	}
+
+	func value(for device: DeviceDetails, deviceInfo: NetworkDevicesInfoResponse? = nil, mainVM: MainScreenViewModel, followState: UserDeviceFollowState?) -> String? {
+		switch self {
+			case .name:
+				return device.name
+			case .gatewayModel:
+				return deviceInfo?.gateway?.model
+			case .bundleName:
+				return device.bundle?.title
+			case .devEUI:
+				return deviceInfo?.weatherStation?.devEui?.convertedDeviceIdentifier ?? device.convertedLabel
+			case .hardwareVersion:
+				return deviceInfo?.weatherStation?.hwVersion
+			case .firmwareVersion:
+				guard let current = device.firmware?.current else {
+					return nil
+				}
+
+				if device.needsUpdate(mainVM: mainVM, followState: followState) {
+					return device.firmware?.versionUpdateString
+				}
+				return current
+			case .lastHotspot:
+				let lastHs = deviceInfo?.weatherStation?.lastHs
+				let timestamp = deviceInfo?.weatherStation?.lastHsActivity?.getFormattedDate(format: .monthLiteralYearDayTime,
+																							 timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+				let elements = [lastHs, timestamp].compactMap { $0 }
+				return elements.isEmpty ? nil : elements.joined(separator: " @ ")
+			case .lastRSSI:
+				return Self.getRssiText(rssi: deviceInfo?.weatherStation?.lastTxRssi,
+										lastActivityDate: deviceInfo?.weatherStation?.lastTxRssiActivity,
+										timezone: device.timezone)
+			case .serialNumber:
+				return deviceInfo?.gateway?.serialNumber?.convertedDeviceIdentifier  ?? device.convertedLabel
+			case .ATECC:
+				return nil
+			case .GPS:
+				guard var gpsSats = deviceInfo?.gateway?.gpsSats else {
+					return nil
+				}
+				gpsSats = LocalizableString.deviceInfoSatellites(gpsSats).localized
+				let timestamp = deviceInfo?.gateway?.gpsSatsLastActivity?.getFormattedDate(format: .monthLiteralYearDayTime,
+																						   timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+				let elements = [gpsSats, timestamp].compactMap { $0 }
+				return elements.joined(separator: " @ ")
+			case .wifiSignal:
+				guard var rssi = deviceInfo?.gateway?.wifiRssi else {
+					return nil
+				}
+				rssi = "\(rssi) \(UnitConstants.DBM)"
+				let timestamp = deviceInfo?.gateway?.wifiRssiLastActivity?.getFormattedDate(format: .monthLiteralYearDayTime,
+																							timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+				let elements = [rssi, timestamp].compactMap { $0 }
+				return elements.joined(separator: " @ ")
+			case .batteryState:
+				return deviceInfo?.weatherStation?.batState?.description
+			case .claimedAt:
+				return deviceInfo?.claimedAt?.getFormattedDate(format: .monthLiteralYearDayTime,
+															   timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+			case .lastGatewayActivity:
+				return deviceInfo?.gateway?.lastActivity?.getFormattedDate(format: .monthLiteralYearDayTime,
+																		   timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+			case .stationModel:
+				return deviceInfo?.weatherStation?.model
+			case .lastStationActivity:
+				return deviceInfo?.weatherStation?.lastActivity?.getFormattedDate(format: .monthLiteralYearDayTime,
+																				  timezone: device.timezone?.toTimezone ?? .current).capitalizedSentence
+			case .stationRssi:
+				return Self.getRssiText(rssi: deviceInfo?.weatherStation?.stationRssi?.formatted(),
+										lastActivityDate: deviceInfo?.weatherStation?.stationRssiLastActivity,
+										timezone: device.timezone)
+		}
+	}
+
+	func warning(for device: DeviceDetails, deviceInfo: NetworkDevicesInfoResponse?) -> (CardWarningConfiguration, VoidCallback)? {
+		switch self {
+			case .name:
+				return nil
+			case .bundleName:
+				return nil
+			case .gatewayModel:
+				return nil
+			case .devEUI:
+				return nil
+			case .hardwareVersion:
+				return nil
+			case .firmwareVersion:
+				return nil
+			case .lastHotspot:
+				return nil
+			case .lastRSSI:
+				return nil
+			case .serialNumber:
+				return nil
+			case .ATECC:
+				return nil
+			case .GPS:
+				return nil
+			case .wifiSignal:
+				return nil
+			case .batteryState:
+				return warningForBatteryState(deviceInfo?.weatherStation?.batState, deviceId: device.id ?? "")
+			case .claimedAt:
+				return nil
+			case .lastGatewayActivity:
+				return nil
+			case .stationModel:
+				return nil
+			case .lastStationActivity:
+				return nil
+			case .stationRssi:
+				return warningForStationRssi(for: deviceInfo?.weatherStation?.stationRssi)
+		}
+	}
+
+	func warningForBatteryState(_ state: BatteryState?, deviceId: String) -> (CardWarningConfiguration, VoidCallback)? {
+		guard let state else {
+			return nil
+		}
+		switch state {
+			case .low:
+				let callback = {
+					WXMAnalytics.shared.trackEvent(.prompt, parameters: [.promptName: .lowBattery,
+																		 .promptType: .warnPromptType,
+																		 .action: .viewAction,
+																		 .itemId: .custom(deviceId)])
+				}
+				return (CardWarningConfiguration(type: .warning,
+												 message: LocalizableString.deviceInfoLowBatteryWarningMarkdown.localized,
+												 closeAction: nil),
+						callback)
+			case .ok:
+				return nil
+		}
+	}
+
+	func warningForStationRssi(for rssi: Int?) -> (CardWarningConfiguration, VoidCallback)? {
+		guard let rssi else {
+			return nil
+		}
+
+		let urlText = "**\(LocalizableString.troubleshootInstructionsHere.localized)**"
+		switch rssi {
+			case _ where rssi >= -80:
+				return nil
+			case _ where rssi >= -95:
+				return (CardWarningConfiguration(type: .warning,
+												 message: LocalizableString.deviceInfoStationRssiWarning.localized,
+												 linkText: LocalizableString.url(urlText, DisplayedLinks.troubleshooting.linkURL).localized,
+												 closeAction: nil),
+						{})
+			default:
+				return (CardWarningConfiguration(type: .error,
+												 message: LocalizableString.deviceInfoStationRssiError.localized,
+												 linkText: LocalizableString.url(urlText, DisplayedLinks.troubleshooting.linkURL).localized,
+												 closeAction: nil),
+						{})
+		}
+	}
+
+	func button(for device: DeviceDetails, mainVM: MainScreenViewModel, followState: UserDeviceFollowState?) -> DeviceInfoButtonInfo? {
+		switch self {
+			case .name:
+				return nil
+			case .bundleName:
+				return nil
+			case .gatewayModel:
+				return nil
+			case .devEUI:
+				return nil
+			case .hardwareVersion:
+				return nil
+			case .firmwareVersion:
+				let buttonInfo = DeviceInfoButtonInfo(icon: .updateFirmwareIcon,
+													  title: LocalizableString.ClaimDevice.updateFirmwareButton.localized,
+													  buttonStyle: .filled())
+				return device.needsUpdate(mainVM: mainVM, followState: followState) ? buttonInfo : nil
+			case .lastHotspot:
+				return nil
+			case .lastRSSI:
+				return nil
+			case .serialNumber:
+				return nil
+			case .ATECC:
+				return nil
+			case .GPS:
+				return nil
+			case .wifiSignal:
+				return nil
+			case .batteryState:
+				return nil
+			case .claimedAt:
+				return nil
+			case .lastGatewayActivity:
+				return nil
+			case .stationModel:
+				return nil
+			case .lastStationActivity:
+				return nil
+			case .stationRssi:
+				return nil
+		}
+	}
+}
